+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for ordering search results by the indexed templates settings. Sort setting needs to include `_indexed_templates` in order for this feature to kick in.
 - Support for pinning specific templates at the top of results list by specifying "pinned_templates" (array) as part of the $args argument for the Finder::find($query, $args) method (or SearchEngine::find($query, $args)).
+- Support for grouping search results by template(s).
+- New QuerySet object for representing queries with multiple sets of database query objects and/or results, such as those resulting from a find operation where results are grouped by templates.
+- New QueryBase base class for Query/QuerySet.
+### Changed
+
+- Query object results and all properties based on results are lazy loaded by default, fetched when first accessed.
 
 ## [0.28.1] - 2021-02-15